Fix default values in array parser

+++ src/parser.h
@@ -39,8 +39,8 @@
 struct parameter {
   char name[PARSER_MAX_LINE_SIZE];
   char value[PARSER_MAX_LINE_SIZE];
-  int used;
-  int is_default;
+  short int used;
+  short int is_default;
 };
 
 struct section {
